+++ src/client/java/io/github/asablock/mdt/command/argument/NoteSubjectArgumentType.java
@@ -1,0 +1,78 @@
+/*
+ * This file is part of mdt. mdt is a client-side mod for Minecraft.
+ * Copyright (C) 2025  asablock
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package io.github.asablock.mdt.command.argument;
+
+import com.mojang.brigadier.StringReader;
+import com.mojang.brigadier.arguments.ArgumentType;
+import com.mojang.brigadier.context.CommandContext;
+import com.mojang.brigadier.exceptions.CommandSyntaxException;
+import com.mojang.brigadier.suggestion.Suggestions;
+import com.mojang.brigadier.suggestion.SuggestionsBuilder;
+import net.minecraft.command.CommandSource;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Locale;
+import java.util.concurrent.CompletableFuture;
+
+public class NoteSubjectArgumentType implements ArgumentType<String> {
+    public static NoteSubjectArgumentType subject() {
+        return new NoteSubjectArgumentType();
+    }
+
+    public static String getSubject(final CommandContext<?> context, final String name) {
+        return context.getArgument(name, String.class);
+    }
+
+    @Override
+    public String parse(StringReader reader) throws CommandSyntaxException {
+        return reader.readString();
+    }
+
+    @Override
+    public String toString() {
+        return "noteSubject()";
+    }
+
+    private static final List<String> EXAMPLES = List.of("unclassified", "Player", "\"foo bar\"");
+
+    @Override
+    public Collection<String> getExamples() {
+        return EXAMPLES;
+    }
+
+    @Override
+    public <S> CompletableFuture<Suggestions> listSuggestions(CommandContext<S> context, SuggestionsBuilder builder) {
+        if (context.getSource() instanceof CommandSource commandSource) {
+            String string = builder.getRemaining().toLowerCase(Locale.ROOT);
+
+            for (String string2 : commandSource.getPlayerNames()) {
+                if (CommandSource.shouldSuggest(string, string2.toLowerCase(Locale.ROOT))) {
+                    builder.suggest(string2);
+                }
+            }
+
+            builder.suggest("unclassified");
+
+            return builder.buildFuture();
+        } else {
+            return builder.suggest("unclassified").buildFuture();
+        }
+    }
+}