+++ src/client/java/io/github/asablock/mdt/toggle/ToggleSerializer.java
@@ -18,20 +18,22 @@
 
 package io.github.asablock.mdt.toggle;
 
-import com.google.gson.*;
+import com.google.gson.JsonIOException;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+import com.google.gson.JsonSyntaxException;
+import io.github.asablock.mdt.Mdt;
 
 import java.io.Reader;
 import java.io.Writer;
 
 public class ToggleSerializer {
-    private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
-
     public static void saveToggles(Writer writer) throws JsonIOException {
         JsonObject root = new JsonObject();
 
         root.add("toggles", Toggles.root.encodeJson());
 
-        GSON.toJson(root, writer);
+        Mdt.PRETTY_PRINTING_GSON.toJson(root, writer);
     }
 
     public static void readToggles(Reader reader) throws JsonIOException, JsonSyntaxException {