+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-# New Features and Improvements
+# New Features
 
-- Added the hotkey Interact to interact with the crosshair target, which is more convenient than using `/minteract crosshairtarget`.
-- Added two hotkeys View Data Components and Print Data Components in Chat to view the data components when hovering over an item stack or print them in chat.
+- Added `/mnote` and `/mnotes` command. Notes are saved in the directory `.minecraft/mdt/notes`.